mcp-oops/pyproject.toml
[project]
name = "mcp-oops"
version = "0.1.0"
description = "MCP-Oops Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncclick>=8.1.8",
    "fastmcp>=2.0.0",
    "mcp>=1.9.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[tool.poetry]
packages = [
    { include = "mcp_oops", from = "src" }
]
[tool.hatch.metadata]
allow-direct-references=true

[project.scripts]
mcp-oops-config = "mcp_oops.main:from_config_cli"
mcp-oops-single = "mcp_oops.main:single_server_cli"


[dependency-groups]
dev = [
  "pytest",
  "pytest-mock",
  "ruff",
]
lint = [
  "ruff"
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.ruff]
extend="../pyproject.toml"

[tool.uv.sources]
fastmcp = { git = "https://github.com/strawgate/fastmcp.git", branch = "proxy_tool" }
fastmcp_agents = { git = "https://github.com/strawgate/fastmcp-agents.git", branch = "main" }
